Add upload_site method for LSST the Docs deploy
upload_site uploads the HTML from render() to LSST the Docs.

The KeeperClient is a neat way to manage the handshake needed for a
build upload. When the client is created it creates a new 'build' with
the Keeper API. As the context manager exits, it tells Keeper that the
build upload is complete.

Adds ltd-conveyor as a dependency.